requirements.txt
pymupdf
pydantic
sentence-transformers
streamlit
python-dotenv
numpy
transformers
torch
pillow
pytesseract
opencv-python
